{cmake} Update three more Standard plugins to targets
- M3C2
- PCV
- SRC
