+++ plugins/core/Standard/qM3C2/CMakeLists.txt
@@ -1,13 +1,13 @@
-option( PLUGIN_STANDARD_QM3C2 "Check to install qM3C2 plugin" OFF )
+option( PLUGIN_STANDARD_QM3C2 "Install qM3C2 plugin" OFF )
 
-if (PLUGIN_STANDARD_QM3C2)
+if( PLUGIN_STANDARD_QM3C2 )
 	project( QM3C2_PLUGIN )
 
-    AddPlugin( NAME ${PROJECT_NAME} )
+	AddPlugin( NAME ${PROJECT_NAME} )
 
-	include_directories( src )
-
+	add_subdirectory( include )
 	add_subdirectory( src )
+	add_subdirectory( ui )
 
-    include( CMakePluginTemplate )
+	include( CMakePluginTemplate )
 endif()
