+++ plugins/core/Standard/qM3C2/include/CMakeLists.txt
@@ -1,0 +1,15 @@
+
+target_sources( ${PROJECT_NAME}
+	PRIVATE
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2.h
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2Commands.h
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2Dialog.h
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2DisclaimerDialog.h
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2Process.h
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2Tools.h
+)
+
+target_include_directories( ${PROJECT_NAME}
+	PRIVATE
+		${CMAKE_CURRENT_SOURCE_DIR}
+)
