+++ plugins/core/Standard/qM3C2/include/qM3C2Tools.h
