+++ plugins/core/Standard/qM3C2/src/CMakeLists.txt
@@ -1,21 +1,9 @@
 
-set( CC_PLUGIN_CUSTOM_HEADER_LIST
-	${CC_PLUGIN_CUSTOM_HEADER_LIST} 
-	${CMAKE_CURRENT_SOURCE_DIR}/qM3C2Commands.h
-	${CMAKE_CURRENT_SOURCE_DIR}/qM3C2Dialog.h
-	${CMAKE_CURRENT_SOURCE_DIR}/qM3C2DisclaimerDialog.h
-	${CMAKE_CURRENT_SOURCE_DIR}/qM3C2Process.h
-	${CMAKE_CURRENT_SOURCE_DIR}/qM3C2Tools.h
-	
-	PARENT_SCOPE
-)
-
-set( CC_PLUGIN_CUSTOM_SOURCE_LIST
-	${CC_PLUGIN_CUSTOM_SOURCE_LIST} 
-	${CMAKE_CURRENT_SOURCE_DIR}/qM3C2Dialog.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/qM3C2DisclaimerDialog.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/qM3C2Process.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/qM3C2Tools.cpp
-
-	PARENT_SCOPE
+target_sources( ${PROJECT_NAME}
+	PRIVATE
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2.cpp
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2Dialog.cpp
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2DisclaimerDialog.cpp
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2Process.cpp
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2Tools.cpp
 )
