+++ plugins/core/Standard/qM3C2/ui/CMakeLists.txt
@@ -1,0 +1,6 @@
+
+target_sources( ${PROJECT_NAME}
+	PRIVATE
+		${CMAKE_CURRENT_LIST_DIR}/disclaimerDlg.ui
+		${CMAKE_CURRENT_LIST_DIR}/qM3C2Dialog.ui
+)
