+++ plugins/core/Standard/qPCV/include/qPCV.h
@@ -30,7 +30,7 @@ class qPCV : public QObject, public ccStdPluginInterface
 {
 	Q_OBJECT
 	Q_INTERFACES(ccStdPluginInterface)
-	Q_PLUGIN_METADATA(IID "cccorp.cloudcompare.plugin.qPCV" FILE "info.json")
+	Q_PLUGIN_METADATA(IID "cccorp.cloudcompare.plugin.qPCV" FILE "../info.json")
 
 public:
 	//! Default constructor
