+++ plugins/core/Standard/qSRA/images/createMapIcon.png
