+++ plugins/core/Standard/qSRA/include/ccMapWindow.h
