+++ plugins/core/Standard/qSRA/include/ccSymbolCloud.h
