+++ plugins/core/Standard/qSRA/include/distanceMapGenerationDlg.h
