+++ plugins/core/Standard/qSRA/src/CMakeLists.txt
@@ -1,0 +1,12 @@
+
+target_sources( ${PROJECT_NAME}
+	PRIVATE
+		${CMAKE_CURRENT_LIST_DIR}/ccSymbolCloud.cpp
+		${CMAKE_CURRENT_LIST_DIR}/distanceMapGenerationDlg.cpp
+		${CMAKE_CURRENT_LIST_DIR}/distanceMapGenerationTool.cpp
+		${CMAKE_CURRENT_LIST_DIR}/dxfProfilesExportDlg.cpp
+		${CMAKE_CURRENT_LIST_DIR}/dxfProfilesExporter.cpp
+		${CMAKE_CURRENT_LIST_DIR}/profileImportDlg.cpp
+		${CMAKE_CURRENT_LIST_DIR}/profileLoader.cpp
+		${CMAKE_CURRENT_LIST_DIR}/qSRA.cpp
+)
