+++ plugins/core/Standard/qSRA/src/ccSymbolCloud.cpp
