+++ plugins/core/Standard/qSRA/ui/CMakeLists.txt
@@ -1,0 +1,7 @@
+
+target_sources( ${PROJECT_NAME}
+	PRIVATE
+		${CMAKE_CURRENT_LIST_DIR}/distanceMapGenerationDlg.ui
+		${CMAKE_CURRENT_LIST_DIR}/dxfProfilesExportDlg.ui
+		${CMAKE_CURRENT_LIST_DIR}/profileImportDlg.ui
+)
